# separate PT and C2 to reduce build time