--- a/benchmarks/operator_benchmark/operator_benchmark.py
+++ b/benchmarks/operator_benchmark/operator_benchmark.py
@@ -2,6 +2,5 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 # TODO (mingzhe09088): get rid of noqa
 import benchmark_runner # noqa
 from benchmark_pytorch import TorchBenchmarkBase # noqa
-from benchmark_caffe2 import Caffe2BenchmarkBase # noqa
 from benchmark_test_generator import * # noqa
 from benchmark_utils import * # noqa